Uninstall deletes directory

+++ packaging/linux_source/uninstall.sh
@@ -1,8 +1,33 @@
 #!/usr/bin/env bash
 
+function confirm() {
+    read -r -p "$1 [y/N] " response
+    case "$response" in
+        [yY][eE][sS]|[yY]) 
+            true
+            ;;
+        *)
+            false
+            ;;
+    esac
+}
+
 DIR=$(realpath $(dirname $0))
 
 pushd "$DIR" > /dev/null
+
+if touch test.txt 2> /dev/null; then
+    rm test.txt
+else
+    echo "Run as root (sudo)."
+    exit 2
+fi
+
+echo "**WARNING:** This will delete the entire directory: $DIR"
+if ! confirm "Continue with uninstall?"; then
+    echo "Cancelling uninstall."
+    exit 0
+fi
 
 echo "Removing virtual environment"
 rm -rf env/
@@ -14,4 +39,5 @@ echo "Removing desktop menu entry"
 xdg-desktop-menu uninstall ArPiRobot-DriveStation.desktop > /dev/null 2>&1
 rm ArPiRobot-DriveStation.desktop
 
-popd > /dev/null
+echo "Deleting $DIR"
+rm -rf "$DIR"
